move character classes into new characterClasses directory #184
