+++ code/act_comm.c
@@ -46,7 +46,7 @@
 #include "spec.h"
 #include "comm.h"
 #include "devextra.h"
-#include "warrior.h"
+#include "characterClasses/warrior.h"
 #include "save.h"
 #include "act_info.h"
 #include "newmem.h"
@@ -60,8 +60,8 @@
 #include "dioextra.h"
 #include "act_move.h"
 #include "handler.h"
-#include "ap.h"
-#include "thief.h"
+#include "characterClasses/ap.h"
+#include "characterClasses/thief.h"
 #include "chardef.h"
 #include "update.h"
 #include "./include/fmt/format.h"
