+++ code/act_obj.c
@@ -49,7 +49,7 @@
 #include "cabal.h"
 #include "devextra.h"
 #include "dioextra.h"
-#include "paladin.h"
+#include "characterClasses/paladin.h"
 #include "fight.h"
 #include "newmem.h"
 #include "comm.h"
@@ -60,7 +60,7 @@
 #include "const.h"
 #include "material.h"
 #include "utility.h"
-#include "warrior.h"
+#include "characterClasses/warrior.h"
 #include "misc.h"
 #include "skills.h"
 #include "./include/fmt/format.h"
