+++ code/characterClasses/class.c
