+++ code/characterClasses/healer.c
@@ -5,17 +5,17 @@
 #include <time.h>
 #include <math.h>
 #include "healer.h"
-#include "handler.h"
-#include "interp.h"
-#include "tables.h"
-#include "recycle.h"
-#include "magic.h"
-#include "db.h"
-#include "lookup.h"
-#include "comm.h"
-#include "devextra.h"
-#include "const.h"
-#include "utility.h"
+#include "../handler.h"
+#include "../interp.h"
+#include "../tables.h"
+#include "../recycle.h"
+#include "../magic.h"
+#include "../db.h"
+#include "../lookup.h"
+#include "../comm.h"
+#include "../devextra.h"
+#include "../const.h"
+#include "../utility.h"
 
 void spell_healing_sleep(int sn, int level, CHAR_DATA *ch, void *vo, int target)
 {
