+++ code/characterClasses/sorcerer.h
@@ -5,7 +5,7 @@
 /****************************************************************************
  *                        Welcome to Sorcerer land.                         *
  ****************************************************************************/
-#include "merc.h"
+#include "../merc.h"
 
 //
 // LOCAL FUNCTIONS
