+++ code/characterClasses/zealot.c
@@ -5,15 +5,15 @@
 #include <time.h>
 #include <math.h>
 #include "zealot.h"
-#include "comm.h"
-#include "handler.h"
-#include "utility.h"
-#include "interp.h"
-#include "tables.h"
-#include "recycle.h"
-#include "magic.h"
-#include "db.h"
-#include "lookup.h"
+#include "../comm.h"
+#include "../handler.h"
+#include "../utility.h"
+#include "../interp.h"
+#include "../tables.h"
+#include "../recycle.h"
+#include "../magic.h"
+#include "../db.h"
+#include "../lookup.h"
 
 void spell_infidels_weight(int sn, int level, CHAR_DATA *ch, void *vo, int target)
 {
