+++ code/characterClasses/zealot.h
@@ -1,7 +1,7 @@
 #ifndef ZEALOT_H
 #define ZEALOT_H
 
-#include "merc.h"
+#include "../merc.h"
 
 //
 // LOCAL FUNCTIONS
