+++ code/const.c
@@ -35,16 +35,16 @@
 #include "magic.h"
 #include "moremagic.h"
 #include "interp.h"
-#include "druid.h"
+#include "characterClasses/druid.h"
 #include "cabal.h"
-#include "healer.h"
-#include "necro.h"
-#include "paladin.h"
+#include "characterClasses/healer.h"
+#include "characterClasses/necro.h"
+#include "characterClasses/paladin.h"
 #include "misc.h"
-#include "ap.h"
-#include "sorcerer.h"
-#include "zealot.h"
-#include "chrono.h"
+#include "characterClasses/ap.h"
+#include "characterClasses/sorcerer.h"
+#include "characterClasses/zealot.h"
+#include "characterClasses/chrono.h"
 #include "fight.h"
 #include "db.h"
