+++ code/fight.c
@@ -51,13 +51,13 @@
 #include "spec.h"
 #include "act_comm.h"
 #include "act_move.h"
-#include "warrior.h"
-#include "sorcerer.h"
+#include "characterClasses/warrior.h"
+#include "characterClasses/sorcerer.h"
 #include "devextra.h"
 #include "skills.h"
-#include "thief.h"
+#include "characterClasses/thief.h"
 #include "save.h"
-#include "paladin.h"
+#include "characterClasses/paladin.h"
 #include "act_obj.h"
 #include "act_info.h"
 #include "newmem.h"
@@ -68,7 +68,7 @@
 #include "dioextra.h"
 #include "db.h"
 #include "misc.h"
-#include "ap.h"
+#include "characterClasses/ap.h"
 #include "iprog.h"
 #include "chardef.h"
 #include "const.h"
