+++ code/interp.c
@@ -70,16 +70,16 @@
 #include "iprog.h"
 #include "const.h"
 #include "utility.h"
-#include "ap.h"
-#include "paladin.h"
+#include "characterClasses/ap.h"
+#include "characterClasses/paladin.h"
 #include "cabal.h"
-#include "necro.h"
+#include "characterClasses/necro.h"
 #include "help.h"
-#include "warrior.h"
+#include "characterClasses/warrior.h"
 #include "misc.h"
 #include "dioextra.h"
-#include "chrono.h"
-#include "thief.h"
+#include "characterClasses/chrono.h"
+#include "characterClasses/thief.h"
 #include "update.h"
 #include "newmem.h"
 #include "lookup.h"
