+++ code/magic.c
@@ -6,15 +6,15 @@
 #include <math.h>
 #include <algorithm>
 #include "magic.h"
-#include "necro.h"
+#include "characterClasses/necro.h"
 #include "recycle.h"
 #include "tables.h"
 #include "act_obj.h"
 #include "act_ente.h"
 #include "act_comm.h"
-#include "warrior.h"
+#include "characterClasses/warrior.h"
 #include "dioextra.h"
-#include "sorcerer.h"
+#include "characterClasses/sorcerer.h"
 #include "act_move.h"
 #include "fight.h"
 #include "act_info.h"
