+++ code/update.c
@@ -49,10 +49,10 @@
 #include "update.h"
 #include "spec.h"
 #include "lookup.h"
-#include "chrono.h"
+#include "characterClasses/chrono.h"
 #include "act_info.h"
 #include "act_move.h"
-#include "thief.h"
+#include "characterClasses/thief.h"
 #include "save.h"
 #include "misc.h"
 #include "act_obj.h"
